requirements.txt
shot-scraper
pyyaml
